+++ src/explorer/models/kafka.ts
@@ -1,5 +1,5 @@
 import { Disposable, TreeItemCollapsibleState } from "vscode";
-import { ClientAccessor } from "../../client";
+import { ClientAccessor, Cluster } from "../../client";
 import { ClusterSettings } from "../../settings";
 import { ClusterItem, NoClusterItem } from "./cluster";
 import { NodeBase } from "./nodeBase";
@@ -16,13 +16,20 @@ export class KafkaModel extends NodeBase implements Disposable {
     }
 
     public async computeChildren(): Promise<NodeBase[]> {
-        const clusters = this.clusterSettings.getAll();
+        const clusters = this.clusterSettings.getAll()
+            .sort(this.sortByNameAscending);
         if (clusters.length === 0) {
             return [new NoClusterItem(this)];
         }
         return clusters.map((c) => {
             return new ClusterItem(this.clientAccessor.get(c.id), c, this);
         });
+    }
+
+    private sortByNameAscending(a: Cluster, b: Cluster): -1 | 0 | 1 {
+        if (a.name.toLowerCase() < b.name.toLowerCase()) { return -1; }
+        if (a.name.toLowerCase() > b.name.toLowerCase()) { return 1; }
+        return 0;
     }
 
     public dispose(): void {
@@ -33,6 +40,6 @@ export class KafkaModel extends NodeBase implements Disposable {
         return this.getChildren()
             .then(clusters =>
                 clusters.find(child => (<ClusterItem>child).cluster.name === clusterName)
-                );
+            );
     }
 }